[Compiler] Unify the ThenFn / FinishOrFn path fragments
Previously, ThenFn and FinishOrFn has unique path descriptors.

After this change, they use the same common @action path descriptor.

While there is some benefit to having them be separate, it was using the
path system for something it wasn't designed for, which is just uniquely
naming points in the computation.