[Compiler] expect() on Context derivations.
Previously, we had the code flat_map and ignore repeated derivations.

This led to subtle bugs where fragments with the same names (which could
arise naturally!) would be dropped uncerimoniously. E.g., if a Contract
declares SomeTraitA::do_thing and SomeTraitB::do_thing, then the
do_thing name would be skipped for A or B whichever is seen second.

Now that we have the renamer, we would end up with do_thing and
do_thing_0 (and if the name do_thing_0 is used separately, do_thing_0_0
etc).

In some contexts, it is natural that we could, if a repeated derivation
is seen, return the Error. However, in other iterator contexts,
returning and processing the error would not be very natural.

It's also the case that the code would be incorrect in an unrecoverable
way were these to not be Ok(), so it is fine to use expect() here.