[Compiler] Custom container wrapper for more efficient unzip.
In the compiler, an unzip is used to simultaneously build up the
continuation points as well as the list of Miniscripts.

In order to build a BTreeMap, one must return an iter of (K, V),
however, we do not want all elements to give a continuation point (TODO:
maybe we do in the future?), so previously we used a dummy entry K/V
that gets repeatedly inserted and removed later.

After this patch we have Option<(K, V)> and custom container wrapper
that filter_maps the iter before collecting.